Common

Common

Components

This section provides a description of the Common components.

Energy

Contains common and simple energy representations. More specific energy structures can be used from other modules such as Amber

ProtoSyn.Common.NullEnergy — Type.

NullEnergy()

Empty placeholder energy container.

Examples

julia> Common.NullEnergy()
Null

source

ProtoSyn.Common.Energy — Type.

Energy(eTotal::Float64)

Simple energy container.

Examples

julia> Common.Energy(1.15)
Energy(eTotal=1.15)

julia> Common.Energy()
Energy(eTotal=0.0)

source

Residue

In ProtoSyn, a Residue object is a collection of atoms (normally an aminoacid) that are identified by a name, a secondary structure (SSTYPE) and are part of a continuous tree of other Residues (have a next Residue).

ProtoSyn.Common.SSTYPE — Type.

SSTYPE

Enum: holds information regarding the secondary structure of each residue in the simulation.

source

ProtoSyn.Common.Residue — Type.

Residue(atoms::Array{Int64, 1}, next::Union{Residue, Int64, Nothing}, name::String, ss::SecondaryStructureType)

Define a residue as part of the system.

Arguments

atoms::Vector{Int64}: list of atom global indices in this residue.

next::Union{Residue, Int64, Nothing}: Next residue in the system, attached to this. Is preferably a Residue instance, but can in certain cases be the index in a Residue list or empty (Nothing).

name::String: Name of the residue. If the residue describes an aminoacid, the correspondent letter is suggested.

ss::SSTYPE: The intitial secondary structure type SSTYPE of this residue.

Examples

julia> Common.Residue([1, 2, 3, 4], Common.Residue([5, 6, 7, 8], nothing, "V", Common.coil), "E", Common.alpha)
Common.Residue(atoms=[1, 2, 3, 4], next=V, name=E, ss=alpha)

source

Dihedral

A Dihedral is a collection of 4 atoms that define a dihedral in the simulated molecule. Many Mutators operate over this dihedrals, changing them in order to explore the conformational space of the system. A Dihedral is part of a Residue and has a defined DIHEDRALTYPE.

ProtoSyn.Common.DIHEDRALTYPE — Type.

DIHEDRALTYPE

Enum: holds information regarding the dihedral type. ```

source

ProtoSyn.Common.Dihedral — Type.

Dihedral(a1::Int64, a2::Int64, a3::Int64, a4::Int64, movable::Array{Int64, 1}, residue::Union{Common.Residue, Int64}, dtype::DIHEDRALTYPE)

Define a dihedral.

Arguments

a1::Int64, a2::Int64, a3::Int64, a4::Int64: global atom indices.

movable::Array{Int64, 1}: List of global atom indices that will be moved during the dihedral movement in this residue.

residue::{Residue}: Residue object that this dihedral belongs to.

dtype::DIHEDRALTYPE: DIHEDRALTYPE.

Examples

julia> Dihedral(2, 3, 5, 7, [5, 6], Common.Residue([1, 2, 3, 4, 5, 6], (...), "A", coil), phi)
Dihedral(a1=2, a2=3, a3=5, a4=7, movable=[5, 6], residue=Common.Residue(atoms=[1, 2, 3, 4, 5, 6], next=V, name=A), type=phi)

See also: Mutators.DihedralMutator

source

ProtoSyn.Common.rotate_dihedral! — Function.

rotate_dihedral!(xyz::Array{Float64, 2}, dihedral::Dihedral, angle::Float64)

Perform a dihedral movement, adding the provided angle (in radians). If the dihedral.dtype is "PHI" or "PSI" the dihedral.residue.next is also rotated and this is propagated recursively until the end of the molecule.

Examples

julia> Mutators.Dihedral.rotate_dihedral!(state.xyz, dihedral, π/2)

See also: Dihedral

source

rotate_dihedral!(xyz::Array{Float64, 2}, a2::Int64, a3::Int64, angle::Float64, dtype::DIHEDRALTYPE, movable::Vector{Int64}[, residue::Union{Residue, Nothing} = nothing])

Base dihedral movement function. Especifies all arguments used in dihedral rotation movement.

Examples

julia> Mutators.Dihedral.rotate_dihedral!(xyz, dihedral.a2, dihedral.a3, π/2, dihedral.dtype, dihedral.movable, dihedral.residue)

See also: Aux.rotation_matrix_from_axis_angle Dihedral Crankshaft

source

Metadata

ProtoSyn Metadata defines additional information of the system that is not necessarily necessary for the basic functions of the library, but allows for a better representation of the system. The AtomMetadata structure holds information related to each atom in the system, such as its element, connections, etc

ProtoSyn.Common.AtomMetadata — Type.

AtomMetadata(name::String[, elem::String = name, res_num::Int64 = 1, res_name::String = "UNK", chain_id::Union{String, Nothing} = nothing, connects::Union{Vector{Int64}, Nothing} = nothing])

Define an atom metadata, containing extra information pertaining the State.

Arguments

name::String: Name of the atom.

elem::String: (Optional) Element of the atom (Default: name).

res_num::Int64: (Optional) Number of the residue this atom belongs to (Default: 1).

res_name::Union{String, Nothing}: (Optional) Name of the residue this atom belongs to (Default: "UNK").

chain_id::String: (Optional) Name of the chain that contains the residue this atom belongs to (Default: nothing).

connects::Union{Vector{Int64}, Nothing}: (Optional) List of global atom indices that this atom is connected to (Default: nothing).

Examples

julia> AtomMetadata("H1", "H", 2, "VAL", "A", [4])
AtomMetadata(name=H1, elem=H, res_num=2, res_name=VAL, chain_id=A, connects=[4])

julia> AtomMetadata("H1")
AtomMetadata(name=H1, elem=H1, res_num=1, res_name=UNK, chain_id=nothing, connects=nothing)

See also: iter

source

ProtoSyn.Common.iter — Function.

iter(data::Vector{AtomMetadata}; property::Symbol = :res_num)

Iterate over an array of AtomMetadata objects, grouping them based on property (Default: :res_num)

Examples

julia> for residue in iter(state.metadata)
    println(residue)
end
[AtomMetadata(name=H1, elem=H1, res_num=1, res_name=UNK, chain_id=nothing, connects=nothing), AtomMetadata(name=H2, elem=H2, res_num=1, res_name=UNK, chain_id=nothing, connects=nothing)]
[AtomMetadata(name=H3, elem=H3, res_num=2, res_name=UNK, chain_id=nothing, connects=nothing), AtomMetadata(name=H4, elem=H4, res_num=2, res_name=UNK, chain_id=nothing, connects=nothing)]

See also: AtomMetadata

source

State

The system state holds information about the current coordinates, energy and forces, aswell as any additional metadata. If iterated over, it returns atom by atom position and metadata.

ProtoSyn.Common.State — Type.

State(size::Int64, energy::AbstractEnergy, xyz::Array{Float64, 2}, forces::Array{Float64, 2}, metadata::Vector{AtomMetadata})

Define the current state of the system, containing the atoms positions, energy and forces applied. If only size::Int64 is provided, an empty State with the given size is created with zeros.

Arguments

size::Int64: Atom count in system.

energy::AbstractEnergy: Current energy of the system (kJ mol⁻¹).

xyz::Array{Float64, 2}: Atom positions in 3 dimensions.

forces::Array{Float64, 2}: Forces applied in each dimension to each atom (kJ mol⁻¹ nm⁻¹)

metadata::Vector{AtomMetadata}: List of atom names.

Examples

julia> Common.State(3)
Common.State(size=3, energy=Null, xyz=[0.0 0.0 0.0; 0.0 0.0 0.0; 0.0 0.0 0.0], forces=[0.0 0.0 0.0; 0.0 0.0 0.0; 0.0 0.0 0.0], metadata=(...))

julia> Common.State(2, Common.NullEnergy(), [1.1 1.1 1.1; 2.2 2.2 2.2], zeros(2, 3), [AtomMetadata(...), AtomMetadata(...), ...])
Common.State(size=2, energy=Null, xyz=[1.1 1.1 1.1; 2.2 2.2 2.2], forces=[0.0 0.0 0.0; 0.0 0.0 0.0], metadata=(...))

See also: AtomMetadata

source

Callback

The CallbackObject allows for independent calls to various functions with individual frequency of output.

ProtoSyn.Common.CallbackObject — Type.

CallbackObject(freq::Int64, callback::Function)

Define the callback function parameters.

Arguments

freq: Frequency (in steps) that the callback function is called.

callback: Actual callback function. This function should have the following signature:

callback(step::Int64, st::Common.State, dr::Drivers.MonteCarlo.MonteCarloDriver, args...)

Examples

julia> Common.CallbackObject(100, Print.as_xyz)
CallbackObject(freq=100, callback=Print.as_xyz)

julia> Common.CallbackObject(Print.as_xyz)
CallbackObject(freq=1, callback=Print.as_xyz)

See also: Print.as_xyz @cbcall

source

Loaders

This section provides a description on how to load a new State, Residue and Dihedral arrays.

ProtoSyn.Common.load_from_gro — Function.

load_from_gro(i_file::String)::Common.State

Return a new Common.State by loading the atom positions and names from the input .gro file. As a default, state.energy is NullEnergy and state.forces are set to zero.

Examples

julia> Common.load_from_gro("molecule.gro")
Common.State(size=2, energy=Null, xyz=[1.1 1.1 1.1; 2.2 2.2 2.2], forces=[0.0 0.0 0.0; 0.0 0.0 0.0], atnames=["C", "O"])

See also: load_from_pdb

source

ProtoSyn.Common.load_from_pdb — Function.

load_from_pdb(i_file::String)::Common.State

Return a new Common.State by loading the atom positions and names from the input .pdb file. As a default, state.energy is NullEnergy and state.forces are set to zero.

Examples

julia> Common.load_from_pdb("molecule.pdb")
Common.State(size=2, energy=Null, xyz=[1.1 1.1 1.1; 2.2 2.2 2.2], forces=[0.0 0.0 0.0; 0.0 0.0 0.0], metadata=(...))

See also: load_from_gro

source

ProtoSyn.Common.load_topology — Function.

load_topology(p::Dict{String, Any})

Parse a dictionary containing the dihedral and residue topology. Return a Dihedral array and a Residue array.

Examples

julia> Mutators.Diehdral.load_topology(p)
(ProtoSyn.Mutators.Dihedral.NewDihedral[...], ProtoSyn.Common.Residue[...])

See also: Aux.read_JSON

source

Conformation Generators

Conformation generators are responsible to change the system State in a defined way.

ProtoSyn.Common.apply_initial_conf! — Function.

apply_initial_conf!(state::State, dihedrals::Vector{Dihedral})

Apply predefined angles to all dihedrals defined in dihedrals, based on the Dihedral.residue.ss, changing the State.xyz to apply the secondary structure. The applied angles (in degrees) are the following:

Beta sheet: PHI = -139.0 | PSI = 135.0 Alpha helix: PHI = -57.0 | PSI = -47.0

Examples

julia> Common.apply_initial_conf(state, dihedrals)

source

Macros

Auxiliary functions that help speed up the system's performance.

ProtoSyn.Common.@cbcall — Macro.

@Common.cbcall callbacks::Tuple{CallbackObject, N} step::Int64 Vararg::Any

(Macro) Call the CallbackObject.function of each CallbackObject in the callbacks Tuple{CallbackObject, N} independently. Each CallbackObject.function is ran depending on the defined CallbackObject.freq and the given step. Vararg holds all the arguments necessary to run the callback function itself.

Examples

julia> @Common.cbcall (callback_object1, callback_object2) 1

source

ProtoSyn.Common.@callback — Macro.

@Common.callback f::Function freq::Int64

(Macro) Create a CallbackObject with the given function f and output frequency freq.

Examples

julia> @Common.callback 10 my_callback1
CallbackObject(freq=10, callback=my_callback1)

julia> @Common.callback my_callback1
CallbackObject(freq=1, callback=my_callback1)

source

ProtoSyn.Common.@faggregator — Macro.

@Common.faggregator name::String f::function Vararg::Any

(Macro) Aggregate multiple functions f in a single variable name. Vararg contains the arguments used by function f

Examples

julia> @faggregator myeval f top1
@faggregator myeval g top2
@faggregator myeval h top3

energy = myevalf(state, false)

source