Graph

ProtoSyn.AbstractDigraph — Type

AbstractDigraph

Supertype for a directed graph node.

ProtoSyn.hasparent — Function

hasparent(c::AbstractDigraph) -> Bool

Test whether the given AbstractDigraph c has a parent.

ProtoSyn.haschildren — Function

haschildren(c::AbstractDigraph) -> Bool

Test whether the given AbstractDigraph c has children.

ProtoSyn.isparent — Function

isparent(p::AbstractDigraph, c::AbstractDigraph) -> Bool

Test whether p is the parent of c.

ProtoSyn.setparent! — Function

setparent!(c::T, p::T) where {T<:AbstractDigraph}

Set p as the parent of c.

ProtoSyn.popparent! — Function

popparent!(c::AbstractDigraph}

Remove the parent from c and make c orphan.

ProtoSyn.popchild! — Function

popchild!(p::AbstractDigraph} -> child

Remove a child from p and return it. The returned element is orphan. Return nothing if the item has no children.

ProtoSyn.adjacency — Function

adjacency(p::AbstractDigraph)

Build and return the adjacency list for all nodes reachable from p.

ProtoSyn.bfs — Function

function bfs(f::Function, c::T) where {T <: AbstractDigraph} -> Nothing

Perform a breadth-first search starting from c and call function f on each visited node.