Types

This section list all core types defined in ProtoSyn.

Builder

Graph

ProtoSyn

Types

Containers

Abstract types

PROTOSYN

Getting Started

Manual

Peptides

Sugars

load stuff

XMLRPC

Note

This is a Note

Containers

All container types are subtypes of AbstractContainer{T} and AbstractDigraph

julia> Atom <: ProtoSyn.AbstractContainer
true

julia> Atom <: ProtoSyn.AbstractDigraph
true

ProtoSyn.Atom — Type

Atom <: AbstractAtom

An Atom type.

Atom(name::String, id::Int, index::Int, symbol::String)

Construct an Atom, with the given name, id, index, and symbol. The created atom has no bonds and container; it is also a free directed-graph node.

Examples

julia> at = Atom("H1", 1, 1, "H")
Atom{/H1:1}

julia> hascontainer(at), isempty(at.bonds)
(false, true)

julia> hasparent(at), haschildren(at)
(false, false)

Abstract types

ProtoSyn.AbstractContainer — Type

AbstractContainer{T} <: AbstractDigraph

Supertype for a container of elements of type T, implementing the AbstractDigraph interface.