Drivers

Drivers

Drivers are a set set of functions that drive the simulation to new states, often exploring the conformational space of the system in Monte Carlo algorithms (with the combination of Mutators with acceptance/rejection rules) or the application of minimizers (such as the Steppest Descent Algorithm). For each Driver a set of required callback functions is further explained, detailing the necessary signature.

Monte Carlo

This section provides a description on the Monte Carlo Driver. This Driver iterates over a set amount of steps (defined in the ConfigParameters), sampling new conformations to the Common.State and accepting or rejecting them based on the Metropolis Algorithm.

ProtoSyn.Drivers.MonteCarlo.ConfigParameters — Type.

ConfigParameters(nsteps::Int64, temperature::Float64)

Define the runtime parameters for the Monte Carlo Driver.

Arguments

nsteps: Total amount of steps to be performed.

temperature: Temperature value to be used in the Metropolis Algorithm.

Examples

julia> Drivers.MonteCarlo.ConfigParameters(100, 0.5)
Drivers.MonteCarlo.ConfigParameters(nsteps=100, temperature=0.5)

See also: load_parameters

source

ProtoSyn.Drivers.MonteCarlo.load_parameters — Function.

load_parameters(p::Dict{String, Any})::ConfigParameters

Load the ConfigParameters from a dictionary.

Examples

julia> Drivers.MonteCarlo.load_parameters(p)
Drivers.MonteCarlo.ConfigParameters(nsteps=100, temperature=0.5)

See also: Aux.read_JSON

source

ProtoSyn.Drivers.MonteCarlo.run! — Function.

run!(state::Common.State, sampler!::Function, evaluator!::Function, params::ConfigParameters[, ostream::IO = stdout, callback::Union{Function, Nothing} = nothing])

Run the main body of the Driver.

Arguments

state::Common.State: Current state of the system to be modified.

sampler!::Function: Responsible for mutating the current state. This function should have the following signature:

sampler!(state::Common.State)

evaluator!::Function: Responsible for evaluating the current state.energy for the Metropolis Algorithm. This function should have the following signature:

evaluator!(state::Common.State, do_forces::Bool)

params::ConfigParameters: Hold the runtime parameters of the Monte Carlo Driver.

ostream::IO: (Optional) Any logging will be written to the supplied ostream.

callback::Union{Function, Nothing}: (Optional) If present, this function will be called if the new conformation was accepted. This function should have the following signature:

callback(state::Common.State, step::Int64)

Tip

The callback function is often a Print function.

Examples

julia> Drivers.MonteCarlo.run(state, my_sampler!, my_evaluator!, params!, callback = my_callback)

See also: Mutators Forcefield.evalenergy!

source

Steepest Descent

This section provides a description on the Steepest Descent Driver. This Driver attempts to minimize the system energy based on the provided Topology, as it calculates the forces acting on each atom according to the defined ConfigParameters. The finalization criteria is:

Maximum number of steps was achieved (ConfigParameters.n_steps).

Maximum force calculated is below the force tolerance (ConfigParameters.f_tol).

ProtoSyn.Drivers.SteepestDescent.ConfigParameters — Type.

ConfigParameters(n_steps::Int64, log_freq::Int64, f_tol::Float64, max_step:Float64)

Define the runtime parameters for the Monte Carlo Driver.

Arguments

n_steps: Total amount of steps to be performed (if convergence is not achieved before).

log_freq: Defines the frequency (in steps) of the logs.

f_tol: Force tolerance. Defines a finalization criteria, as the steepest descent is considered converged if the maximum force calculated is below this value.

max_step: Defines the maximum value ɣ that the system can jump when applying the forces.

Examples

julia> Drivers.SteepestDescent.ConfigParameters(100, 5, 1e-3, 0.1)
Drivers.SteepestDescent.ConfigParameters(n_steps=100, log_freq=5, f_tol=1e-3, max_step=0.1)

See also: load_parameters

source

ProtoSyn.Drivers.SteepestDescent.load_parameters — Function.

load_parameters(p::Dict{String, Any})::ConfigParameters

Load the ConfigParameters from a dictionary.

Examples

julia> Drivers.SteepestDescent.load_parameters(p)
Drivers.SteepestDescent.ConfigParameters(n_steps=100, log_freq=5, f_tol=1e-3, max_step=0.1)

See also: Aux.read_JSON

source

ProtoSyn.Drivers.SteepestDescent.run! — Function.

run!(state::Common.State, evaluator!::Function, params::ConfigParameters[, ostream::IO = stdout, callback::Union{Function, Nothing} = nothing])

Run the main body of the Driver.

Arguments

state::Common.State: Current state of the system to be modified.

evaluator!::Function: Responsible for evaluating the current state.energy and calculate the resulting forces. This function should have the following signature:

evaluator!(state::Common.State, do_forces::Bool)

params::ConfigParameters: Hold the runtime parameters of the Steepest Descent Driver.

ostream::IO: (Optional) Any logging will be written to the supplied ostream.

callback::Union{Function, Nothing}: (Optional) If present, this function will be called every ConfigParameters.log_freq. This function should have the following signature:

callback(state::Common.State, step::Int64)

Tip

The callback function is often a Print function.

Examples

julia> Drivers.SteepestDescent.run(state, my_evaluator!, params!, callback = my_callback)

See also: Forcefield.evalenergy!

source