Forcefield

Force-field

Components

This section provides a description on all the individual components that comprise the Forcefield.

ProtoSyn.Forcefield.HarmonicBond — Type.

HarmonicBond(a1::Int64, a2::Int64, k::Float64, b0::Float64)

Harmonic Bond of the form

\[E(r_{ab}) = \frac{1}{2}k_{ab}(r_{ab} - b_0)^2\]

where

\[r_{ab} = \|\vec{r_{ab}}\| = \|\vec{r_b} - \vec{r_a}\|\]

Arguments

a1::Int64, a2::Int64: global atom indices.

k::Float64: force constant (kJ mol⁻¹ nm⁻²).

b0::Float64: equilibrium bond length (nm).

Examples

julia> Forcefield.HarmonicBond(1, 2, 2500, 0.19)
Forcefield.HarmonicBond(a1=1, a2=2, k=2500.0, b0=0.19)

source

ProtoSyn.Forcefield.HarmonicAngle — Type.

HarmonicAngle(a1::Int64, a2::Int64, a3::Int64, k::Float64, θ::Float64)

Harmonic Angle of the form

\[E(θ_{abc})=\frac{1}{2}k_{abc}(\theta_{abc}-\theta)^{2}\]

Arguments

a1::Int64, a2::Int64, a3::Int64: global atom indices.

k::Float64: force constant (kJ mol⁻¹ rad⁻²).

θ::Float64: equilibrium angle value (rad).

Examples

julia> Forcefield.HarmonicAngle(1, 2, 3, 670.0, 1.92)
Forcefield.HarmonicAngle(a1=1, a2=2, a3=3, k=670.0, θ=1.92)

source

ProtoSyn.Forcefield.DihedralCos — Type.

DihedralCos(a1::Int64, a2::Int64, a3::Int64, a4::Int64, k::Float64, θ::Float64, mult::Float64)

Periodic Dihedral of the form

\[E(\phi_{abcd})=K_{\phi}(1+cos(n\phi_{abcd}-\phi))\]

Arguments

a1::Int64, a2::Int64, a3::Int64, a4::Int64: global atom indices.

k::Float64: force constant (kJ mol⁻¹).

θ::Float64: equilibrium angle value (deg).

mult::Float64: multiplicity.

Examples

julia> Forcefield.DihedralCos(1, 2, 3, 4, 10.46, 180.0, 2.0)
Forcefield.DihedralCos(a1=1, a2=2, a3=3, a4=4, k=10.46, θ=180.0, mult=2.0)

source

ProtoSyn.Forcefield.Atom — Type.

Atom(name::String, σ::Float64, ϵ::Float64, q::Float64, excls::Array{Int64, 1}, pairs::Array{Int64, 1})

Define an atom. σ, ϵ and q describe the non-bonded interactions between atoms:

The Lennard-Jones interaction is in the form:

\[E(r_{ab}) = 4ϵ_{ab}\left ( (\frac{σ_{ab}}{r_{ab}})^{12}-(\frac{σ_{ab}}{r_{ab}})^{6}\right )\]

where the Lorentz-Berthelot rule is applied. σ is the arithmetic average and ϵ is the geometric average:

\[σ_{ab}=\frac{σ_{a}+σ_{b}}{2}\]

\[ϵ_{ab}=\sqrt{(ϵ_{a}ϵ_{b})}\]

For this reason, σ and ϵ are applied here in the reduced form: $\frac{σ}{2}$ and $\sqrt{ϵ}$.

The Coulomb interation is in the form:

\[E(r_{ab})=k_{ϵ}\frac{q_{a}q_{b}}{r_{ab}^{2}}\]

where

\[k_{ϵ}=\frac{1}{4πϵ_{0}}=138.935485\,kJ\,nm\,mol⁻¹\,e⁻¹\]

For this reason, q is applied here in the reduced form: $q\times \sqrt{k_{ϵ}}$

Exclusion list contains all atom indices who are excluded from non-bonded interactions (i.e. are at 3 or less connections from this atom - includes pairs). Pair list contains atoms that are at 3 connections from this atom, and are involved in 1-4 interactions (and have a different combination rule as a result).

Arguments

name::String: Atom name (example: "C", "H", etc).

σ::Float64: half the finite distance at which the inter-particle potential is zero (nm).

ϵ::Float64: square root of the depth of the potential well (kJ mol⁻¹).

q::Float64: adjusted atom charge (eletron).

excls::Array{Int64, 1}: exclusion list (as global atom indices).

pairs::Array{Int64, 1}: pair list containing atoms that interfere in 1-4 interations (as global atom indices)

Examples

julia> Forcefield.Atom("N", 0.325, 0.711, 0.0017, [0, 1, 2, 3, 4, 5], [4, 5])
Forcefield.Atom(name="N", σ=0.325, ϵ=0.711, q=0.0017, excls=[0, 1, 2, 3, 4, 5], pairs=[4, 5])

source

ProtoSyn.Forcefield.Topology — Type.

Topology(atoms::Array{Atom}, bonds::Array{HarmonicBond}, angles::Array{HarmonicAngle}, dihedralsCos::Array{DihedralCos})

Gather all topology components.

Arguments

atoms::Array{Atoms}

bonds::Array{HarmonicBond}

angles::Array{HarmonicAngle}

dihedralsCos::Array{DihedralCos}

Examples

julia> Forcefield.Forcefield(atoms, bonds, angles, dihedrals)
Forcefield.Topology(
 atoms=ProtoSyn.Forcefield.Atom[Forcefield.Atom(name="N", σ=0.325, ϵ=0.711, q=0.0017, excls=[0, 1, 2, 3, 4, 5], pairs=[4, 5]), ...],
 bonds=ProtoSyn.Forcefield.HarmonicBond[Forcefield.HarmonicBond(a1=1, a2=2, k=2500.0, b0=0.19), ...],
 angles=ProtoSyn.Forcefield.HarmonicAngle[Forcefield.HarmonicAngle(a1=1, a2=2, a3=3, k=670.0, θ=1.92), ...],
 dihedralsCos=ProtoSyn.Forcefield.DihedralCos[Forcefield.DihedralCos(a1=1, a2=2, a3=3, a4=4, k=10.46, θ=180.0, mult=2.0), ...])

source

ProtoSyn.Forcefield.Energy — Type.

DihedralCos(eBond::Float64, eAngle::Float64, eDihedral::Float64, eLJ::Float64, eLJ14::Float64, eCoulomb::Float64, eCoulomb14::Float64, eTotal::Float64)

Energy components.

Arguments

a1::Int64, a2::Int64, a3::Int64, a4::Int64: global atom indices.

k::Float64: force constant (kJ mol⁻¹).

θ::Float64: equilibrium angle value (deg).

mult::Float64: multiplicity.

Examples

julia> Forcefield.Energy()
Forcefield.Energy(
 eBond=0.0,
 eAngle=0.0,
 eDihedral=0.0,
 eLJ=0.0,
 eLJ14=0.0,
 eCoulomb=0.0,
 eCoulomb14=0.0,
 eTotal=0.0
)

julia> Forcefield.Energy(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8)
Forcefield.Energy(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8)

source

Evaluators

This section provides a description on how to use the previously documented Components to calculate their energy.

ProtoSyn.Forcefield.evalbond! — Function.

evalbond!(bonds::Array{Forcefield.HarmonicBond}, state::Common.State, do_forces::Bool = false) -> Float64

Evaluate an array of Bonds using the given state.

source