Mutators

Mutators

Mutators are modules that introduce a single conformational change on the system, according to a set of runtime parameters. These can be combined into multiple steps, defining a Driver.

Dihedral

This section provides a description on the Dihedral mutator, responsible for performing a single dihedral movement according to a set of ConfigParameters.

ProtoSyn.Mutators.Dihedral.ConfigParameters — Type.

ConfigParameters(p_mut::Float64 = 0.1)

Define the runtime parameters for Dihedral movements.

Arguments

p_mut::Float64: Probability of mutation of this dihedral (Default: 0.1).

Examples

julia> Mutators.Diehdral.ConfigParameters(0.2)
ConfigParameters(p_mut=0.2)

julia> Mutators.Diehdral.ConfigParameters()
ConfigParameters(p_mut=0.1)

source

ProtoSyn.Mutators.Dihedral.NewDihedral — Type.

NewDihedral(a1::Int64, a2::Int64, a3::Int64, a4::Int64, movable::Array{Int64, 1}, residue::Union{Common.Residue, Int64}, dtype::String)

Define a dihedral.

Arguments

a1::Int64, a2::Int64, a3::Int64, a4::Int64: global atom indices.

movable::Array{Int64, 1}: List of global atom indices that will be moved during the dihedral movement in this residue.

residue::Union{Common.Residue, Int64}: Residue that this dihedral belongs to. Should be a Common.Residue object.

dtype::String: Dihedral type (i.e. "PHI", "PSI", ...)

Examples

julia> Mutators.Diehdral.NewDihedral(2, 3, 5, 7, [5, 6], Common.Residue([1, 2, 3, 4, 5, 6], (...), "A"), "PSI")
Dihedral(a1=2, a2=3, a3=5, a4=7, movable=[5, 6], residue=Common.Residue(atoms=[1, 2, 3, 4, 5, 6], next=V, name=A), type="PSI")

See also: load_topology

source

ProtoSyn.Mutators.Dihedral.load_topology — Function.

load_topology(p::Dict{String, Any})

Parse a dictionary containing the dihedral and residue topology. Return a NewDihedral array and a Common.Residue array.

Examples

julia> Mutators.Diehdral.load_topology(p)
(ProtoSyn.Mutators.Dihedral.NewDihedral[...], ProtoSyn.Common.Residue[...])

See also: Aux.read_JSON

source

ProtoSyn.Mutators.Dihedral.run! — Function.

run!(state::Common.State, dihedrals::Array{NewDihedral, 1}, params::Config.Parameters, angle_sampler::Function[, ostream::IO = stdout])

Iterate over a list of NewDihedral (dihedrals) and perform dihedral movements on the current Common.State. The probability of each dihedral undergo movements is defined in the ConfigParameters (params.pmut). The new angle is obtained from angle_sampler, who should return a Float64 in radians. After movement, the Common.State is updated with the new conformation. Any logging is written to ostream (Default: stdout).

Examples

julia> Mutators.Diehdral.run!(state, dihedrals, params, () -> randn())

See also: rotate_dihedral!

source

ProtoSyn.Mutators.Dihedral.rotate_dihedral! — Function.

rotate_dihedral!(xyz::Array{Float64, 2}, dihedral::NewDihedral, angle::Float64)

Perform a dihedral movement, adding the provided angle (in radians). If the dihedral.dtype is "PHI" or "PSI" the dihedral.residue.next is also rotated and this is propagated recursively until the end of the molecule.

Examples

julia> Mutators.Diehdral.rotate_dihedral(state.xyz, dihedral, π/2)

See also: run! Aux.rotation_matrix_from_axis_angle

source