Peptides

ProtoSyn.Peptides.isproline — Function

isproline(r::Residue) -> Bool

Determine if a residue is a proline.

ProtoSyn.Peptides.build_graph! — Function

build_graph!(top::Topology)

Utility function for building the atom and residue graphs of peptides. The root atom of the atom graph for each segment is taken as the N of the first residue of each segment.

ProtoSyn.Peptides.grammar — Function

grammar([T=Float64,] dir::AbstractString=resource_dir)

Build a LGrammar for peptides, taking as variables the fragments in dir. The peptidejoin function is included as the default operator. The returned L-grammar is required for building peptides from fragments.

Examples

julia> g = Peptides.grammar();
julia> pose = Builder.build(grammar, "AAGASTASSE")
...

!!! NOTE requires update

ProtoSyn.Peptides.setss! — Function

setss!(pose::Pose, (ϕ, ψ, ω)::NTuple{3,Number})

Set the ϕ, ψ and ω backbone angles of all residues in the given pose. This function is usefull for setting the secondary structure of a pose. This function acts on the internal coordinates and does not update cartesian coordinates, although a request for conversion is made. It is up to the calling function/user to explicitly synchornize coordinates via sync!.

ProtoSyn.Peptides.grammar — Method

grammar([T=Float64,] dir::AbstractString=resource_dir)

Build a LGrammar for peptides, taking as variables the fragments in dir. The peptidejoin function is included as the default operator. The returned L-grammar is required for building peptides from fragments.

Examples

julia> g = Peptides.grammar();
julia> pose = Builder.build(grammar, "AAGASTASSE")
...

!!! NOTE requires update

ProtoSyn.Peptides.isproline — Method

isproline(r::Residue) -> Bool

Determine if a residue is a proline.

ProtoSyn.Peptides.setss! — Method

setss!(pose::Pose, (ϕ, ψ, ω)::NTuple{3,Number})

Set the ϕ, ψ and ω backbone angles of all residues in the given pose. This function is usefull for setting the secondary structure of a pose. This function acts on the internal coordinates and does not update cartesian coordinates, although a request for conversion is made. It is up to the calling function/user to explicitly synchornize coordinates via sync!.

ProtoSyn.Peptides.build_graph! — Method

build_graph!(top::Topology)

Utility function for building the atom and residue graphs of peptides. The root atom of the atom graph for each segment is taken as the N of the first residue of each segment.