Common

The Common module is a pecial type of module, in ProtoSyn. Its function is to group commonly performed tasks, such as defining a default EnergyFunction instances or defining commonly used Callback instances.

Default Energy Function

ProtoSyn.Common.default_energy_function — Function

default_energy_function(::Type{T}) where {T <: AbstractFloat}
default_energy_function()

Returns a default energy function for ProtoSyn.

Note:

If no Type{T} is provided, will use ProtoSyn.Units.defaultFloat;

Examples

julia> ProtoSyn.Common.default_energy_function()
🗲  Energy Function (4 components):
+----------------------------------------------------------------------+
| Index | Component name                                | Weight (α)   |
+----------------------------------------------------------------------+
| 1     | TorchANI_ML_Model                             |      1.000   |
| 2     | Caterpillar_Solvation                         |      0.010   |
| 3     | Bond_Distance_Restraint                       |      1.000   |
| 4     | Cα-Cα_Clash_Restraint                         |    100.000   |
+----------------------------------------------------------------------+

ProtoSyn.Common.default_xmlrpc_energy_function — Function

default_xmlrpc_energy_function(::Type{T}) where {T <: AbstractFloat}
default_xmlrpc_energy_function()

Returns a default energy function for ProtoSyn, where a XML-RPC version of each component is employed (when available).

Note:

If no Type{T} is provided, will use ProtoSyn.Units.defaultFloat;

Note:

Since this energy function employs the XML-RPC protocol whenever possible, it is slower, but safe in term of CUDA running out of memory.

See also

default_energy_function

Examples

julia> ProtoSyn.Common.default_xmlrpc_energy_function()
🗲  Energy Function (4 components):
+----------------------------------------------------------------------+
| Index | Component name                                | Weight (α)   |
+----------------------------------------------------------------------+
| 1     | TorchANI_ML_Model_XMLRPC                      |      1.000   |
| 2     | Caterpillar_Solvation                         |      0.010   |
| 3     | Bond_Distance_Restraint                       |      1.000   |
| 4     | Cα-Cα_Clash_Restraint                         |    100.000   |
+----------------------------------------------------------------------+

Default Callbacks

ProtoSyn.Common.default_energy_step_callback — Function

default_energy_step_callback(n::Int)::Callback

Returns a simple Callback that prints the current step and total energy value of the Pose every n steps.

Examples

julia> ProtoSyn.Common.default_energy_step_callback(100)
✉  Callback:
+----------------------------------------------------------------------+
| Index | Field                     | Value                            |
+----------------------------------------------------------------------+
| 1     | Event                     | energy_step                      |
| 2     | Frequency                 | 100                              |
+----------------------------------------------------------------------+

ProtoSyn.Common.default_energy_step_frame_callback — Function

default_energy_step_frame_callback(n::Int, filename::String)::Callback

Returns a simple Callback that prints the current step and total energy value of the Pose every n steps, while printing the current structure to the given filename (in PDB format).

Examples

julia> ProtoSyn.Common.default_energy_step_frame_callback(100, "test.pdb")
✉  Callback:
+----------------------------------------------------------------------+
| Index | Field                     | Value                            |
+----------------------------------------------------------------------+
| 1     | Event                     | energy_step_frame                |
| 2     | Frequency                 | 100                              |
+----------------------------------------------------------------------+