Radius of gyration

The radius of gyration of a molecule is, in short, a measure of how "elongated" that structure is in each of the 3 dimensions (x, y and z). This can be useful, for example, in globular systems where an elongation in any of the dimensions is usually not likely to be a stabilizing conformational change. The following methods allow ProtoSyn to calculate the radius of gyration of a Pose and attribute a tentative energy value.

ProtoSyn.Calculators.RG.calc_radius_gyration — Function

calcradiusgyration(pose::Pose, [selection::Opt{AbstractSelection} = nothing])

Calculate each dimension (X, Y and Z) radius of gyration (the root mean square deviation to all of the structure's Atom instances from the Pose pose center_of_mass). Each Atom type mass is retrieved from ProtoSyn.Units.mass.

See also

calc_radius_gyration_energy get_default_rg

Examples

julia> ProtoSyn.Calculators.SASA.calc_radius_gyration(pose)
3×1 Matrix{Float64}:
4.976766149665673
5.501503499039599
10.74091335737219

ProtoSyn.Calculators.RG.calc_radius_gyration_energy — Function

calc_radius_gyration_energy(pose::Pose, selection::Opt{AbstractSelection}, update_forces::Bool)

Calculate the sum of all dimensions (X, Y and Z) radius of gyration. This Calculator does not calculate forces. As such, update_forces has no effect and exists only in order to standardize calls between Calculators.

See also

calc_radius_gyration get_default_rg

Examples

julia> ProtoSyn.Calculators.SASA.calc_radius_gyration_energy(pose, nothing, false)
(21.21918300607746, nothing)

ProtoSyn.Calculators.RG.get_default_rg — Function

get_default_rg(;[α::T = 1.0]) where {T <: AbstractFloat}

Return the default Radius of Gyration energy EnergyFunctionComponent. α sets the component weight (on an EnergyFunction instance, 1.0 by default). This function employs calc_radius_gyration_energy as the :calc function.

See also

calc_radius_gyration_energy

Examples

julia> ProtoSyn.Calculators.SASA.get_default_sasa_energy()
🞧  Energy Function Component:
+---------------------------------------------------+
| Name           | Radius_Gyration                  |
| Alpha (α)      | 1.0                              |
| Update forces  | false                            |
| Calculator     | calc_radius_gyration_energy      |
+---------------------------------------------------+
└──  ○  Selection: nothing