Electrostatics

The Peptides module Electrostatics section expands the Core module methods (See Electrostatics) specifically to assign default charges on protein and peptide sequences.

ProtoSyn.Peptides.Calculators.Electrostatics.assign_default_charges! — Function

assign_default_charges!(pose::Pose, [res_lib::LGrammar = Peptides.grammar], [selection::Opt{AbstractSelection}]; [supress_warn::Bool = false])

Assign default charges to Pose pose from the given LGrammar res_lib entry, by Atom name. If an AbstractSelection selection is provided, only apply charges to the selected Atom instances. For non-canonical aminoacids and ligands (any Residue without an entry on ProtoSyn.three_2_one dictionary) and any Residue whose template have different Atom names, a warning is shown. Set supress_warn to true to ignore these warnings (false, by default).

Note:

This method is an overload of assign_default_charges!. Check it's documentation page for more information. In comparison, this method is specialized for Peptides (automatically identifies terminal atoms and adds correct charges their caps).

See also

assign_acc2_eem_charges_from_file!

Examples

julia> ProtoSyn.Peptides.Calculators.Electrostatics.assign_default_charges!(pose)
1143-element Vector{Float64}:
 -0.025115728872692304
 -0.025115728872692304
 -0.025115728872692304
 (...)