Input and Output (IO)

The following methods allow the user to load and export models and data from and to files, and build on top of the Core methods (See Input and Output (IO)).

ProtoSyn.Peptides.load — Function

load([::Type{T}], filename::AbstractString; bonds_by_distance::Bool = false) where {T <: AbstractFloat}

Load the given filename into a Pose, parametrized by T. If this is not provided, the default ProtoSyn.Units.defaultFloat is used instead. The file format is infered from the extension (Supported: .pdb, .yml). If bonds_by_distance is set to true (false, by default), the CONECT records will be complemented with bonds infered by distance. The threshold distances for each pair of atoms is defined in ProtoSyn.bond_lengths. Infers parenthood and ascedence from bonds (N-Residue instances have the connected C-Residue as child).

Note:

This function is an extension of ProtoSyn.load.

Examples

julia> ProtoSyn.Peptides.load("1ctf.pdb", bonds_by_distance = true)
Pose{Topology}(Topology{/2a3d:61708}, State{Float64}:
 Size: 1140
 i2c: false | c2i: false
 Energy: Dict(:Total => Inf)
)