State

This section lists functions that work on the system State, as expansions to the Core methods (See State). These methods are subdivided by topics, for organization purposes:

Applying a Secondary Structure

Applying a Secondary Structure

Making use of the Secondary Structure types (in specific, the SecondaryStructureTemplate instance), the setss! method allows a user to quickly apply a given Secondary Structure to a Pose or a subset of selected Residue instances.

ProtoSyn.Peptides.setss! — Function

setss!(state::State, ss::SecondaryStructureTemplate, residues::Vector{Residue})

Set the phi ϕ, psi ψ and omega ω backbone angles of all Residue instances in the given residues vector to match the provided SecondaryStructureTemplate. This function acts on the internal coordinates and does not update cartesian coordinates (using the sync! method), although a request for conversion is made (by calling the request_i2c!).

setss!(pose::Pose, ss::SecondaryStructureTemplate)
setss!(pose::Pose, ss::SecondaryStructureTemplate, sele::ProtoSyn.AbstractSelection)
setss!(pose::Pose, ss::SecondaryStructureTemplate, residue::Residue)

Set the phi ϕ, psi ψ and omega ω backbone angles of all Residue instances in the given Pose pose to match the provided SecondaryStructureTemplate. If an optional AbstractSelection sele is provided, apply the transformation only to the selected Residue instances.

See also

SecondaryStructure

Examples

julia> ProtoSyn.Peptides.setss!(pose, ProtoSyn.Peptides.SecondaryStructure[:helix])
State{Float64}:
 Size: 343
 i2c: true | c2i: false
 Energy: Dict(:Total => Inf)