Aux

Aux

This section provides a description of miscellaneous auxiliary functions.

ProtoSyn.Aux.read_JSON — Function.

read_JSON(i_file::String)

Read a JSON file and return a dictionary with the content.

Examples

julia> Aux.read_JSON("i_file.json")
Dict{Any, Any}()

source

ProtoSyn.Aux.rotation_matrix_from_axis_angle — Function.

rotation_matrix_from_axis_angle(axis::Vector{Float64}, angle::Float64)

Return a rotation matrix based on the provided axis and angle (in radians).

Examples

julia> Aux.rotation_matrix_from_axis_angle([1.1, 2.2, 3.3], π/2)
3×3 Array{Float64,2}:
  0.0714286  -0.658927  0.748808
  0.944641    0.285714  0.16131 
 -0.320237    0.695833  0.642857

See also: rotate_dihedral!

source

ProtoSyn.Aux.calc_dih_angle — Function.

calc_dih_angle(a1::Vector{Float64}, a2::Vector{Float64}, a3::Vector{Float64}, a4::Vector{Float64})

Calculates the dihedral angle produced between a1, a2, a3 and a4, in radians.

Examples

julia> Aux.calc_dih_angle([1.0, 1.0, 1.0], [2.1, 2.1, 2.1], [3.0, 2.0, 5.0], [5.0, 5.0, 5.0])
3.141592653589793

See also: apply_initial_conf!

source