Mutators

Mutators

Mutators are modules that introduce a single conformational change on the system, according to a set of runtime parameters. These can be combined into multiple steps, defining a Driver.

Dihedral

This section provides a description on the Dihedral mutator, responsible for performing a single dihedral movement according to a DihedralMutator set of parameters.

ProtoSyn.Mutators.Dihedral.DihedralMutator — Type.

DihedralMutator(dihedrals::Vector{Common.Dihedral}, angle_sampler::Function, p_mut::Float64, step_size::Float64)

Holds all necessary parameters for the correct simulation of dihedral movements.

Arguments

dihedrals::Vector{Common.Dihedral}: List of dihedrals avaliable to be rotated in dihedral movements.

angle_sampler::Function: Function responsible for defining the new angle for the dihedral. Should return a Float64.

p_mut::Float64: Probability of rotation of each dihedral.

step_size::Float64: Scalar that defines the amount of change resulting for a dihedral movement.

Examples

julia> Mutators.Diehdral.DihedralMutator(dihedrals, 0.05, randn, 0.25)
DihedralMutator(dihedrals=68, p_pmut=0.05, angle_sampler=randn, step_size=0.25)

See also: run!

source

ProtoSyn.Mutators.Dihedral.run! — Function.

run!(state::Common.State, mutator::DihedralMutator)

Iterate over a list of Common.Dihedral (dihedrals) and perform dihedral movements on the current Common.State. The probability of each dihedral undergo movements is defined in the DihedralMutator.p_mut. The new angle is obtained from DihedralMutator.angle_sampler, who should return a Float64 in radians. After movement, the Common.State is updated with the new conformation.

Examples

julia> Mutators.Dihedral.run!(state, mutator)

See also: Common.rotate_dihedral!

source

Crankshaft

A Crankshaft movement is a rotation of all atoms between two randomly chosen alpha carbons according to a set of CrankshaftMutator parameters.

ProtoSyn.Mutators.Crankshaft.CrankshaftMutator — Type.

CrankshaftMutator(dihedrals::Vector{Common.Dihedral}, angle_sampler::Function, p_mut::Float64, step_size::Float64)

Holds all necessary parameters for the correct simulation of crankshaft movements.

Arguments

dihedrals::Vector{Common.Dihedral}: List of dihedrals avaliable to be rotated in crankshaft movements.

angle_sampler::Function: Function responsible for defining the rotation angle. Should return a Float64.

p_mut::Float64: Probability of rotation of each pair of alpha carbons.

step_size::Float64: Scalar that defines the amount of change resulting for a crankshaft movement.

Examples

julia> Mutators.Crankshaft.CrankshaftMutator(dihedrals, randn, 0.05, 0.25)
CrankshaftMutator(dihedrals=68, angle_sampler=randn, p_pmut=0.05, step_size=0.25)

julia> Mutators.Crankshaft.CrankshaftMutator(dihedrals, randn)
CrankshaftMutator(dihedrals=68, angle_sampler=randn, p_pmut=0.0, step_size=0.0)

See also: run!

source

ProtoSyn.Mutators.Crankshaft.run! — Function.

run!(state::Common.State, mutator::CrankshaftMutator)

Iterate over a list of Common.Dihedral (dihedrals) and perform crankshaft movements on the current Common.State. The probability of each pair of alpha carbons undergo movement is defined in the CrankshaftMutator.p_mut. The new angle is obtained from CrankshaftMutator.angle_sampler, who should return a Float64 in radians. After movement, the Common.State is updated with the new conformation.

Examples

julia> Mutators.Crankshaft.run!(state, mutator)

See also: rotate_crankshaft!

source

ProtoSyn.Mutators.Crankshaft.rotate_crankshaft! — Function.

rotate_crankshaft!(xyz::Array{Float64, 2}, dihedral1::Common.Dihedral, dihedral2::Common.Dihedral, angle::Float64)

Perform a crankshaft movement, adding the provided angle (in radians) to all atoms between the alpha carbon in dihedral1 and dihedral2. The last residue side_chain is also rotated and should be relaxed back to a equilibrium state (See: SteepestDescent).

Examples

julia> Mutators.Dihedral.rotate_crankshaft!(state.xyz, dihedral1, dihedral2, π/2)

source