Common

Common

Components

This section provides a description of the Common components, such as Common.State.

ProtoSyn.Common.NullEnergy — Type.

NullEnergy()

Empty placeholder energy container.

Examples

julia> Common.NullEnergy()
Null

source

ProtoSyn.Common.State — Type.

State(size::Int64, energy::AbstractEnergy, xyz::Array{Float64, 2}, forces::Array{Float64, 2}, atnames::Array{String, 1})

Define the current state of the system, containing the atoms positions, energy and forces applied. If only size::Int64 is provided, an empty State with the given size is created with zeros.

Arguments

size::Int64: Atom count in system.

energy::AbstractEnergy: Current energy of the system (kJ mol⁻¹).

xyz::Array{Float64, 2}: Atom positions in 3 dimensions.

forces::Array{Float64, 2}: Forces applied in each dimension to each atom (kJ mol⁻¹ nm⁻¹)

atnames::Array{String, 1}: List of atom names.

Examples

julia> Common.State(3)
Common.State(size=3, energy=Null, xyz=[0.0 0.0 0.0; 0.0 0.0 0.0; 0.0 0.0 0.0], forces=[0.0 0.0 0.0; 0.0 0.0 0.0; 0.0 0.0 0.0], atnames=String[])

julia> Common.State(2, Common.NullEnergy(), [1.1 1.1 1.1; 2.2 2.2 2.2], zeros(2, 3), ["C", "O"])
Common.State(size=2, energy=Null, xyz=[1.1 1.1 1.1; 2.2 2.2 2.2], forces=[0.0 0.0 0.0; 0.0 0.0 0.0], atnames=["C", "O"])

source

ProtoSyn.Common.Residue — Type.

Residue(atoms::Array{Int64, 1}, next::Union{Residue, Int64, Nothing}, name::String)

Define a residue as part of the system.

Arguments

atoms::Array{Int64, 1}: list of atom global indices in this residue.

next::Union{Residue, Int64, Nothing}: Next residue in the system, attached to this. Is preferably a Residue instance, but can in certain cases be the index in a Residue list or empty (Nothing).

name::String: Name of the residue. If the residue describes an aminoacid, the correspondent letter is suggested.

Examples

julia> Common.Residue([1, 2, 3, 4], Common.Residue([5, 6, 7, 8], nothing, "V"), "E")
Common.Residue(atoms=[1, 2, 3, 4], next=V, name=E)

source

Loader

This section provides a description on how to load a new Common.State.

ProtoSyn.Common.load_from_gro — Function.

load_from_gro(i_file::String)::Common.State

Return a new Common.State by loading the atom positions and names from the input .gro file. As a default, state.energy is NullEnergy and state.forces are set to zero.

Examples

julia> Common.load_from_gro("molecule.gro")
Common.State(size=2, energy=Null, xyz=[1.1 1.1 1.1; 2.2 2.2 2.2], forces=[0.0 0.0 0.0; 0.0 0.0 0.0], atnames=["C", "O"])

See also: load_from_pdb

source

ProtoSyn.Common.load_from_pdb — Function.

load_from_pdb(i_file::String)::Common.State

Return a new Common.State by loading the atom positions and names from the input .pdb file. As a default, state.energy is NullEnergy and state.forces are set to zero.

Examples

julia> Common.load_from_pdb("molecule.pdb")
Common.State(size=2, energy=Null, xyz=[1.1 1.1 1.1; 2.2 2.2 2.2], forces=[0.0 0.0 0.0; 0.0 0.0 0.0], atnames=["C", "O"])

See also: load_from_gro

source