Drivers

Drivers

Drivers are a set set of functions that drive the simulation to new states, often exploring the conformational space of the system in Monte Carlo algorithms (with the combination of Mutators with acceptance/rejection rules) or the application of minimizers (such as the Steppest Descent Algorithm).

Monte Carlo

This section provides a description on the Monte Carlo Driver. This Driver iterates over a set amount of n_steps (defined in the MonteCarloDriver), sampling new conformations to the Common.State and accepting or rejecting them based on the Metropolis Algorithm.

ProtoSyn.Drivers.MonteCarlo.MonteCarloDriver — Type.

MonteCarloDriver(sampler!::Function, evaluator!::Function, [, temperature::Float64 = 1.0, n_steps::Int64 = 0])

Define the runtime parameters for the Monte Carlo simulation. No sampler! movement is performed by default, since n_steps = 0.

Arguments

sampler!::Function: Responsible for generating a new structure to be evaluated. This function should have the following signature:

sampler!(state::Common.State)

evaluator!::Function: Responsible for evaluating the system energy. This function should have the following signature:

evaluator!(state::Common.State, do_forces::Bool)

temperature::Float64: (Optional) Temperature of the system, determines acceptance in the Metropolis algorithm (Default: 1.0).

n_steps: (Optional) Total amount of steps to be performed (Default: 0).

Examples

julia> Drivers.MonteCarlo.MonteCarloDriver(my_sampler!, my_evaluator!, 10.0, 1000)
MonteCarloDriver(sampler=my_sampler!, evaluator=my_evaluator!, temperature=10.0, n_steps=1000)

julia> Drivers.MonteCarlo.MonteCarloDriver(my_sampler!, my_evaluator!)
MonteCarloDriver(sampler=my_sampler!, evaluator=my_evaluator!, temperature=1.0, n_steps=0)

Tip

Both my_sampler! and my_evaluator! functions often contain pre-defined function avaliable in Mutators and Forcefield modules, respectively.

See also: run!

source

ProtoSyn.Drivers.MonteCarlo.run! — Function.

run!(state::Common.State, driver::MonteCarloDriver[, callbacks::Tuple{Common.CallbackObject}...])

Run the main body of the driver. Creates a new conformation based on driver.sampler!, evaluates the new conformation energy using driver.evaluator!, accepting it or not depending on the driver.temperature in a Metropolis algorithm. This Monte Carlo process is repeated for driver.n_steps, saving the accepted structures to state and calling all the callbacks.

Examples

julia> Drivers.MonteCarlo.run!(state, driver, my_callback1, my_callback2, my_callback3)

source

Steepest Descent

This section provides a description on the Steepest Descent Driver. This Driver attempts to minimize the system energy based on the provided Topology, as it calculates the forces acting on each atom according to the defined SteepestDescentDriver. The finalization criteria is:

Maximum number of steps was achieved (SteepestDescentDriver.n_steps).

Maximum force calculated is below the force tolerance (SteepestDescentDriver.f_tol).

The gamma (γ) applied in the next step of the Steepest Descent is below machine precision.

ProtoSyn.Drivers.SteepestDescent.SteepestDescentDriver — Type.

SteepestDescentDriver(evaluator!::Function[, n_steps::Int64 = 0, f_tol::Float64 = 1e-3, max_step:Float64 = 0.1])

Define the runtime parameters for the Steepest Descent simulation. If n_steps is zero, a single point energy calculation is performed.

Arguments

evaluator!::Function: Responsible for evaluating the current state.energy and calculate the resulting forces. This function should have the following signature:

evaluator!(state::Common.State, do_forces::Bool)

n_steps: (Optional) Total amount of steps to be performed (if convergence is not achieved before) (Default: 0).

f_tol: (Optional) Force tolerance. Defines a finalization criteria, as the steepest descent is considered converged if the maximum force calculated is below this value (Default = 1e-3).

max_step: (Optional) Defines the maximum value ɣ that the system can jump when applying the forces (Default: 0.1).

ostream: (Optional) Defines the output stream for logging

Examples

julia> Drivers.SteepestDescent.SteepestDescentDriver(my_evaluator!, 100, 1e-3, 0.1)
SteepestDescentDriver(evaluator=my_evaluator!, n_steps=100, f_tol=1e-3, max_step=0.1)

julia> Drivers.SteepestDescent.SteepestDescentDriver(my_evaluator!, f_tol = 1e-6)
SteepestDescentDriver(evaluator=my_evaluator!, n_steps=0, f_tol=1e-6, max_step=0.1)

Tip

The my_evaluator! function often contains an aggregation of pre-defined functions avaliable in Forcefield. It is possible to combine such functions using the @faggregator macro.

See also: Amber.evaluate! run!

source

ProtoSyn.Drivers.SteepestDescent.run! — Function.

run!(state::Common.State, driver::SteepestDescentDriver[, callback::Union{Common.CallbackObject, Nothing} = nothing])

Run the main body of the Driver. If driver.n_steps is zero, a single point energy calculation is performed.

Arguments

state::Common.State: Current state of the system to be modified.

driver::SteepestDescentDriver: Defines the parameters for the SteepestDescent simulation. See SteepestDescentDriver.

callbacks::Vararg{Common.CallbackObject, N}: (Optional) Tuple of CallbackObjects (Default: empty).

Tip

The callback function often contains a Print function.

Examples

julia> Drivers.SteepestDescent.run(state, steepest_descent_driver, callback1, callback2, callback3)

source